+++ bounded-contexts/budget/src/test/scala/works/iterative/incubator/budget/ui/transaction_import/TransactionImportPresenterLiveSpec.scala
@@ -1,9 +1,8 @@
-package works.iterative.incubator.budget.application.service
+package works.iterative.incubator.budget.ui.transaction_import
 
 import works.iterative.incubator.budget.domain.model.AccountId
 import works.iterative.incubator.budget.domain.service.*
 import works.iterative.incubator.budget.infrastructure.adapter.*
-import works.iterative.incubator.budget.ui.transaction_import.TransactionImportPresenter
 import works.iterative.incubator.budget.ui.transaction_import.models.*
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
